+++ src-ui/images.ts
@@ -20,6 +20,7 @@ const imageSrcs = [
   { name: "monsters4", src: "/images/tiles/monsters4.png" },
   { name: "monsters5", src: "/images/tiles/monsters5.png" },
   { name: "monstersbig2", src: "/images/tiles/monstersbig2.png" },
+  { name: "monstersbig3", src: "/images/tiles/monstersbig3.png" },
   { name: "monstersbig4", src: "/images/tiles/monstersbig4.png" },
   { name: "monstersbig5", src: "/images/tiles/monstersbig5.png" },
   { name: "icesmallbg", src: "/images/tiles/icesmallbg.png" },