+++ CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Black Market Rooms/Chunks
   - Haunted Castle Rooms/Chunks
   - Ice Caves Rooms/Chunks
+  - Wet Fur Rooms/Chunks
   - Display Room Type on room views
+  - Support Room group notes that apply to all rooms in a grouping
+  - Add support for 79 length level to display bugged levels
 
 ### Fixed